Add image in the capirotada.html file and adjust
I created a new directory called "images" to store the pictures for the
website. Then I added the picture for the first recipe as an jpg image.

I added the link of the image inside the body of the capirotada.html
file by adding the code <img src="capirotada.jpg" alt="Images of Mexican
Capirotada">.

Since the image was too large for the webpage, I adjusted the size by
using width and height.