Add description and lists under the image
Added a description under the previously imported image in the
capirotada.html file. I used an <h3> for the heading of the description
and <p> element for the paragraph with the description text under the
heading

I added a <h3> heading for the ingredients, followed by an uordered lists
using the element "ul". Then I added another <h3> heading and an ordered
list using "ol" element for the steps of the recipe.