Add a new file with another recipe
I added a new html file called "sushi.html" inside the recipes folder.
I added a link in the index.html file with the code: <a href="recipes/
sushi.html"Homemade Sushi</a>. I put this link and the previous link
to the capirotada.html file inside an uordered list using "ul" element.

I added a new image inside the body of the sushi.html file by puting a
link to a new image from the "images" folder inside odin-recipes.

I followed the same code pattern as the capirotada.html file to add the
description and the uordered and ordered lists with the ingredients and
the steps of the recipe.